backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.9"
dependencies = [
    "aiosqlite>=0.22.1",
    "fastapi>=0.128.8",
    "httpx>=0.28.1",
    "openai>=2.31.0",
    "playwright>=1.40",
    "python-dotenv>=1.2.1",
    "sqlalchemy>=2.0.49",
    "uvicorn>=0.39.0",
]

[project.optional-dependencies]
test = ["pytest>=8.0", "pytest-asyncio>=0.24"]
